Restore column width (except when "fit to width" option is enabled)

+++ src/main/java/org/jabref/gui/importer/ImportInspectionDialog.java
@@ -582,8 +582,8 @@ public class ImportInspectionDialog extends JabRefDialog implements ImportInspec
         }
 
         for (int i = 0; i < INSPECTION_FIELDS.size(); i++) {
-            int width = InternalBibtexFields.getFieldLength(INSPECTION_FIELDS.get(i));
-            glTable.getColumnModel().getColumn(i + PAD).setPreferredWidth(width);
+            Double width = InternalBibtexFields.getFieldLength(INSPECTION_FIELDS.get(i));
+            glTable.getColumnModel().getColumn(i + PAD).setPreferredWidth(width.intValue());
         }
     }
